Add a session filter
This filter adds a cookie with a session uuid and an expiration time
when a request comes in which doesn't have session information.

Any ulterior request will read the session information from the cookie
(or from a header if the session was provided by an uphill system).

If the session is not expired (`now` is before `expirationDate`) the
filter will postpone the current expiration date to
`now() + sessionDuration` otherwise it will assign a new session to the
request.

* session uuid key is configurable
* session duration is configurable